+++ Changelog.md
@@ -20,6 +20,8 @@
   - Each actor in the combat tracker on that faction's side will be examined for a possible contradiction, and a chat card will prompt with a dice roll for each actor. The 0, 1 or 1-4 cases are based on none, one or both of the following being true:
     - The actor's cosm does not match either of the cosms of the current scene.
     - Any item on the has a higher `Tech Level` (if set) or higher `secondary axiom` than the highest axioms of the cosms of the current scene.
+- Provide automation for detecting **disconnection** during skill use.
+- When **disconnected**, any items on the Actor which exceed the current scene's axioms will have a line drawn through the item's name.
 - Drama Card Line tooltip shows if current scene is `Standard` or `Dramatic`
 - Report a **fragile item breaking** on a Mishap involving its use.